+++ public/scripts/index.js
@@ -86,7 +86,9 @@ socket.on(MESSAGE_TYPE.SERVER_MESSAGE, async (data) => {
       );
       drawBoughtSquares();
       await forceClientsToRefresh();
-      blunderBotSay(`@${data.user} You're on the board! Let's gooooo!`);
+      blunderBotSay(
+        `@${data.user} You've got the ${square} square! Let's gooooo!`
+      );
     }
   } else if (data.type === MESSAGE_TYPE.REFRESH) {
     load();